+++ install-deps.sh
@@ -355,12 +355,19 @@ else
 	if $with_jaeger; then
 	    build_profiles+=",pkg.ceph.jaeger"
 	fi
+
+	wget -qO - https://dist.apache.org/repos/dist/dev/arrow/KEYS | $SUDO apt-key add -
+	echo "deb [arch=amd64] https://apache.jfrog.io/artifactory/arrow/ubuntu $(lsb_release -sc) main" | $SUDO tee /etc/apt/sources.list.d/arrow.list
+	$SUDO apt update
+
 	$SUDO env DEBIAN_FRONTEND=noninteractive mk-build-deps \
 	      --build-profiles "${build_profiles#,}" \
 	      --install --remove \
 	      --tool="apt-get -y --no-install-recommends $backports" $control || exit 1
 	$SUDO env DEBIAN_FRONTEND=noninteractive apt-get -y remove ceph-build-deps
 	if [ "$control" != "debian/control" ] ; then rm $control; fi
+	$SUDO rm -f /etc/apt/sources.list.d/arrow.list
+
         ;;
     centos|fedora|rhel|ol|virtuozzo)
         builddepcmd="dnf -y builddep --allowerasing"
