+++ src/rgw/CMakeLists.txt
@@ -2,6 +2,14 @@ find_program(GPERF gperf)
 if(NOT GPERF)
   message(FATAL_ERROR "Can't find gperf")
 endif()
+
+find_package(Arrow QUIET)
+if(Arrow_FOUND)
+  set(ARROW_LIBRARIES "-larrow -lparquet")
+  add_definitions(-D_ARROW_EXIST)
+  message("-- arrow is installed, radosgw/s3select-op is able to process parquet objects")
+endif()
+
 function(gperf_generate input output)
   add_custom_command(
     OUTPUT ${output}
@@ -122,6 +130,7 @@ set(librgw_common_srcs
   rgw_rest_realm.cc
   rgw_rest_role.cc
   rgw_rest_s3.cc
+  rgw_s3select.cc
   rgw_role.cc
   rgw_sal.cc
   rgw_sal_rados.cc
@@ -204,6 +213,7 @@ target_link_libraries(rgw_common
     ${CURL_LIBRARIES}
     ${EXPAT_LIBRARIES}
     ${LUA_LIBRARIES}
+    ${ARROW_LIBRARIES}
   PUBLIC
     spawn)
 target_include_directories(rgw_common
@@ -302,6 +312,7 @@ target_link_libraries(rgw_a
     common_utf8 global
     ${CRYPTO_LIBS}
     ${LUA_LIBRARIES}
+    ${ARROW_LIBRARIES}
     OATH::OATH
   PUBLIC
     rgw_common
