+++ src/rgw/rgw_s3select.h
@@ -1,0 +1,8 @@
+// -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
+// vim: ts=8 sw=2 smarttab ft=cpp
+//
+
+namespace rgw::s3select {
+RGWOp* create_s3select_op();
+}
+
